feat(git): shared RepoConnectForm + clickable commits; center icon tiles
- Extract the repo-connect form into components/git/RepoConnectForm.jsx
  (provider strip + URL/branch/auto-deploy + optional tracked paths +
  connected state) — the canonical reusable surface. components/wordpress/
  GitConnectForm.jsx is now a thin WordPress-flavored wrapper over it, so
  WordPress renders identically and other pages can drop the shared form in.
- Make the WordPress commit log clickable: each SHA links to the commit on
  the remote (GitHub/GitLab /commit, Bitbucket /commits; https + scp-ssh).
- Fix off-center icon tiles on /services/new (grid + place-items + block svg
  instead of inline-flex).
